[better_errors] Port the Pallas debug info mechanisms to the new JAX TracingDebugInfo.
Now that we carry debug informatiion in Jaxpr we can remove the Pallas-specific
tracking of the `func_src_info`: we remove `NameAndSrcInfo` and its usage.

Had to make some changes to ensure that the kernel and the index_map Jaxprs have
the debug info; this fixed some tests.

We will be able to simplify some `assert debug_info is not None` once we
make more changes throughout Jax to ensure that the debugging info is propagated.