[Mosaic GPU] Maintain a 1:1 correspondence between `Barrier_ref` and `MosaicGPU_Barrier`.
Before this change, the barrier in the mlir dialect was just a pointer. It was missing information about offsets and parities. To make those usable and enable future enhancements (e.g. barrier_wait and barrier_await ops in the dialect) we need the dialect barrier to maintain the full state.

PiperOrigin-RevId: 720932225