Replace setup.py with pyproject.toml
This allows package managers to see all of the dependencies statically
rather than having to download Jax and run setup.py to do the analysis.

This allows uv to build a virtual environment automatically using:
`uv venv`, and a lock file with `uv lock -U`.

It also adds a dev-dependency optional group, which makes development a
bit easier.  For example, with uv, the venv for development can be
updated with: `uv sync --extra dev` or `uv sync --all-extras`.

This changes the workflow slightly: Instead of editing setup.py upon
release, edit fix_pyproject.py, and run it to update pyproject.toml.