Fix assertion fails on DebugInfo with JAX_USE_DIRECT_LINEARIZE=1
When using JAX_USE_DIRECT_LINEARIZE=1 there was an assertion that
result_paths is not None. This is because in this case we create
a DebugInfo that is not used with `lu.wrap_init` (which normally
takes care of the result_paths). To allow others to make
progress, I just fill in a (wrong) result_paths, and a TODO
to come back and fix this later.