Fix persistent iteration
We accomplish this by:

1. Round-robin iterating through input buffer slots based on
`persistent_ki`, which tracks k-loop iterations over the entire
lifetime of a CTA worker, and not just a single work item.

2. Arriving on an ab_empty_barrier even in the last iteration of a
k-loop, so that barrier phase is flipped correctly across work items.

3. Waiting on an ab_empty_barrier in all iterations of all work items'
k-loops, except for the first `max_concurrent_steps` iterations of the
very first work item's k-loop.