[attrs] allow setattr on a previously non-existant attr
Before this change, we handled attrs for initial-style primitives like jit/scan
like this:
1. the traceable would form a jaxpr and see what attrs were touched (by
   jax_getattr or jax_setattr),
2. for each such attr, the traceable would do jax_getattr to get the current
   value, tree-flatten, pass the flat valuesinto the (pure) bind, get the new
   values out, tree-unflatten, then jax_setattr the result.

That approach would error if the function called `jax_setattr` to set a
previously non-existant attr. That is, this would work:

```python
from jax.experimental.attrs import jax_setattr
class Thing: ...
thing = Thing()
jax_setattr(thing, 'x', 1.0)
```
but it wouldn't work under a `jax.jit`.

This commit makes the same code work under a jit. We just
1. in partial_eval.py's `to_jaxpr`, ensure attrs added during jaxpr formation
   are deleted, using a special sentinel value `dne_sentinel` to indicate the
   attribute initially did not exist before tracing;
2. in pjit.py's `_get_states`, when reading initial attr values before the
   pjit_p bind, if the attribute does not exist we don't try to read it and
   instead just use `dne_sentinel` as the value, which is a convenient empty
   pytree;
3. in pjit.py's `_attr_token` for jit caching, when forming the cache key based
   on the current attr states, we map attrs that don't exist to `dne_sentinel`
   (rather than just erroring when the attr doesn't exist, as before).

In short, we use a special value to indicate "does not exist".

If `jax_getattr` supported the 'default' argument, the code would be a little
cleaner since we could avoid the `if hasattr` stuff. And that's probably a
useful feature to have anyway. We can add that in a follow-up.

This PR only makes setattr-to-nonexistant-attr work with jit. We'll add scan
etc in follow-ups.